ref: move vercel testutils to fixtures
